+++ src/events.rs
@@ -1,6 +1,7 @@
 use crate::game::file::Status;
 
 pub enum Event {
+    AnimationTick,
     StatusUpdate(Status),
     Exit,
 }